requirements.txt
Flask
pyyaml
jsonschema
jsonformatter
xsdata
